backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
alembic
python-dotenv
pydantic
passlib[bcrypt]
python-jose[cryptography]
psycopg2-binary
pandas
python-multipart
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
